Move han zi to components
